# Use fst crate for improved performance