--- a/process_corpus/Cargo.toml
+++ b/process_corpus/Cargo.toml
@@ -7,3 +7,4 @@ edition = "2018"
 
 [dependencies]
 anyhow = "1.0"
+fst = "0.4"