--- a/search_corpus/Cargo.toml
+++ b/search_corpus/Cargo.toml
@@ -9,6 +9,9 @@ edition = "2018"
 cgi = "0.6"
 json = "0.12"
 url = "2.1"
+fst = "0.4"
+regex-automata = { version = "0.1.9", features = ["transducer"] }
+memmap = "0.7"
 
 [dev-dependencies]
 criterion = "0.3"