After investigating the [fst crate](https://github.com/BurntSushi/fst) suggested in #1, I was able to get much improved performance in some cases.

These benchmarks were taken on my desktop computer:

&nbsp;|without FST|with FST|
--|--|--|
small_word_many_options| 39 ms | 7 ms (-81%)|
small_word_few_options| 37 ms | 199 **u**s (-99%!) |
long_word_many_options| 53 ms | 72 ms (+34%) |
long_word_few_options| 38 ms | 351 **u**s (-99%!) |

So the FST implementation is much faster until there are 6 or more question marks in the pattern, at which point it becomes slower. This implementation uses a hybrid approach - if there are less than 6 question marks, it uses the FST, otherwise is does the old regular-expression matching on the word list.