@Eh2406 2021-09-25
I know I am getting nerd sniped at this point, but... 

If the performance is proportional to the number of `?` then it is probably spending its time in `dense::Builder`, according to the [readme](https://github.com/BurntSushi/regex-automata):

> While the regex crate doesn't necessarily sport fast compilation times, the regexes in this crate are almost universally slow to compile, especially when they contain large Unicode character classes. For example, on my system, compiling `\w{3}` with byte classes enabled takes just over 1 second and almost 5MB of memory! (Compiling a sparse regex takes about the same time but only uses about 500KB of memory.) Conversly, compiling the same regex without Unicode support, e.g., `(?-u)\w{3}`, takes under 1 millisecond and less than 5KB of memory. For this reason, you should only use Unicode character classes if you absolutely need them!

So I wonder what the effect of adding `(?-u)` to ``build_regex` would be as `absent_letter_builder` can match any unicode char. Alternatively, `absent_letter_builder` can start as `"[a-z]&&[^-'".to_string();` instead of `"[^-'".to_string();` to make for a even more targeted regex.