@gregstoll 2021-09-25
Ooh, interesting! I played around with this and after some fighting got it to work - see commit [ce4ee210](https://github.com/gregstoll/wheeloffortune/commit/ce4ee210bef5f99b1a49b67d0cda48fcfb8224be). Unfortunately it doesn't seem to be any faster, for the regex or for the FST. Worth a try though!